backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
pyodbc==5.0.1
python-dotenv==1.0.0
SQLAlchemy==2.0.23